Make extension_renamer reusable
